frontend/Quizz_App_For_Studying_With_UI/requirements.txt
Flask
requests
Flask-SQLAlchemy
sympy
google-generativeai
gunicorn
